% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pals.R
\name{pv_styles}
\alias{pv_styles}
\title{Extract Styles}
\usage{
pv_styles(...)
}
\arguments{
\item{...}{A list of color names}
}
\value{
A list of hex codes
}
\description{
An internal function to extract the named colors
}
