% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pals.R
\name{scale_color_spartan}
\alias{scale_color_spartan}
\alias{scale_fill_spartan}
\title{Spartan Colors Palette for ggplot2}
\usage{
scale_color_spartan(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_spartan(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a PVCSD color palette (see Details)}
}
\description{
Set the colors and fill for ggplot objects to colors that fit the spartan palettes
}
\details{
\subsection{Palettes}{
\itemize{
\item \code{main}  The default palette scales from blue to grey
\item \code{blues} A monochromatic blue palette
\item \code{mono}  A single color palette
\item \code{cat}   A color palette for categorical variables
}
}
}
