% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pals.R
\name{spartan_palette}
\alias{spartan_palette}
\title{Fetch the spartan palettes}
\usage{
spartan_palette(palette = "main", reverse = FALSE, ...)
}
\value{
A palette
}
\description{
An internal function that creates a color ramp palette if needed
}
