% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pals.R
\docType{data}
\name{spartanPalettes}
\alias{spartanPalettes}
\title{A list of PVCSD color palettes}
\format{An object of class \code{list} of length 4.}
\usage{
spartanPalettes
}
\value{
A list of named colors
}
\description{
An Internal list of palettes for consistant styling
}
\keyword{datasets}
