% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pals.R
\docType{data}
\name{spartanStyles}
\alias{spartanStyles}
\title{Spartan Colors and Styles}
\format{An object of class \code{character} of length 13.}
\usage{
spartanStyles
}
\description{
An internal named list of colors for a consistant theme for PVCSD
}
\keyword{datasets}
